Make Future as Promise lazy
If we are rejecting a promise whenever a future is rejected,
that's no good because it causes the runtime to complain that a
rejected promise is never handled.

This makes the promise creation lazy, that way we never even create
the promise unless someone is trying to use a future as a promise.